+++ test/fixtures/classes/product/Product.php
@@ -1,0 +1,13 @@
+<?php
+namespace hoge\fuga\product;
+
+class Product {
+    private Name $name;
+    private Price $price;
+}
+class Name {
+    private string $name;
+}
+class Price {
+    private int $price;
+}